Update package to (almost) run on example.fasta
Most of the updates of Kmerfeatures.py to a working Snakemake workflow and command-line interface have been implemented. The CLI is working, and the dryrun properly builds the DAG for the pipeline. The workflow is stuck at rule preprocess when saving outputs; the dtypes must be properly resolved before this step can continue, and utils.save_hdf() and utils.load_hdf() are in-progress to properly write and read the output of rule preprocess in the near future.