Change aggregate_kmers input to labels instead of features
Fixes a minor bug discovered with different alphabets/kmer sizes where the full feature label is not output to the feature files, but is contained in the labels file. Because the kmer ID parser (kmerfeatures.utils.get_output_kmers) relies on the full KMER-RED{N}-xxxxx identifier, the kmer identifiers were not parsed correctly, and the kmer IDs were not being collected in the aggregation step. However, fixing the feature file output is still planned in the future.