Subtract background family scores from kmer probability score
Background family scores are now subtracted from family scores when the family is present in a specified background file in the config.yaml. Has been tested to work locally, but relies on family name to be present in the filename of the background file (e.g. with one large background file containing multiple families, does not work yet.). Will test with a larger background set later.