Account for class weights in model (partially addresses #15 bug)
The model currently performs worse than chance despite updates to the normalization method. One potential cause is that the models may be predicting only one class, and thus incorrectly predicts nearly all of the positive examples. To account for this, `class_weight='balanced'` has been added as a model parameter.