Increase default file creation wait time (`--latency-wait`) to 30s
Default snakemake latency wait time (i.e. the amount of time the filesystem waits for output files to be created) is 3 seconds, which leads to errors in instances where the output files are not created quickly enough due to latency issues. The default wait time has been increased to minimize these errors.