Modify CLI to explicitly read mode variable. Update documentation
Small documentation formatting (making sure the correct special character was being used to represent file directory branches) changes were made.

The CLI now reads the execution mode as a variable so that it can be read into the code in the future if needed. To accommodate this change, the example config must also contain a `mode` variable and that has been updated accordingly.