Move k-fold splitting to `score` step. Minor reformatting
Some Pythonic reformatting (via black and snakefmt) has been applied to some files.

The k-fold splitting for cross-validation has been moved to the `rules.model.score` step. The indices of the train/test splits for each fold are then saved in the dataframe for future use in the model step. Partially addresses issues raised in #39 and #41, though with incomplete implementation.