[debug] Explicitly check for alphabet=None. Fixes #50
Previously, `alphabet=None` was not explicitly checked as an alphabet and thus did not function as the other alphabets. This has been resolved.

Minor changes to the cluster workflow (changing string formatting to f-strings, and checking for param == "None" vs. param is None to account for erroneous reading of params as strings from config file)