Terminate workflow via error instead of printing to console
Changes the behavior of the `min_filter` parameter when the filter is set too high; previously, an error message was printed to console, but the snakemake workflow will continue running until the missing input files are detected in the next rule. This has been changed to terminate with an error immediately instead. Thus, the workflow will not attempt to run subsequent rules that are invalid past the `min_filter` step.