[fix] Resolve `AttributeError` when no config file specified
Snekmer now has the following behavior for finding config files:
(1) If a config.yaml file is specified on the command line, but `--directory` is not, Snekmer assumes a file called config.yaml exists in th
e current working directory.
(2) If a config.yaml file is specified and a working directory is also specified with the `--directory` flag, Snekmer assumes that config.ya
ml file has been specified as a full path (i.e. not relative to the `--directory` specification). In other words, the `--directory` value is not used to modify the path to the config.yaml file.
(3) If no config.yaml file is specified but a working directory is specified with the `--directory` flag, Snekmer assumes that there is a config.yaml file contained within the specified working directory and will search for the path (i.e. `snekmer {mode} --directory /path/to/workdir` -> Snekmer assumes a config file at `/path/to/workdir/config.yaml`)
(4) If neither a config.yaml file nor a `--directory` are specified on the command line, Snekmer assumes that a config.yaml file exists within the current working directory.